Removed run_tasks and moved connection_object to module_utils
